[Kashif]Refactored:Createsd a separate jar and include the jar in the Census Analyser Project
