+++ src/main/java/opencsv/OpenCSVBuilder.java
@@ -1,4 +1,4 @@
-package com.cg;
+package opencsv;
 
 import java.io.Reader;
 import java.util.Iterator;
